Fixes and initial test suite for media parts of the iPod module and tests for the Audio class
The media test suite for the iPod module is in a working state, but the iPod fails at the tests.
Also a few fixes were done to the iPod module, some because of the tests. Now an iPod module does not create the database as soon as it is initiliazed. This is much better, but more improvements are needed in the locking / db cache code.
A test suite was created for the Audio class. It is very simple, covering only the basics.